refactor: clean up project organization and move interfaces to Abstractions
- Remove 10 stale .gitkeep files from directories with actual content
- Remove empty placeholder directories (Enrichers/, ActivitySources/)
- Move IOperationScope and IOperationScopeFactory to Abstractions/ folder
- Update namespace to HVO.Enterprise.Telemetry.Abstractions
- Add using directives to all 14 referencing files (src, tests, benchmarks)
