--- src/HVO.Common/Extensions/.gitkeep
@@ -1,1 +1,0 @@
-# Extension methods will go here
