+++ src/HVO.Enterprise.Telemetry/OperationScopeExtensions.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using HVO.Enterprise.Telemetry.Abstractions;
 
 namespace HVO.Enterprise.Telemetry
 {
